Fix types, format code, move check for parameter approach into choose.py
This simplifies the Connection object and fully moves _all_ parameter logic
into the parameters module.
