+++ src/databricks/sql/utils.py
@@ -320,8 +320,6 @@ class RequestErrorInfo(
         return user_friendly_error_message
 
 
-
-
 def create_arrow_table_from_arrow_file(file_bytes: bytes, description) -> pyarrow.Table:
     arrow_table = convert_arrow_based_file_to_arrow_table(file_bytes)
     return convert_decimals_in_arrow_table(arrow_table, description)
