+++ tests/unit/test_param_escaper.py
@@ -3,7 +3,7 @@ import unittest, pytest, decimal
 from typing import Any, Dict
 from databricks.sql.parameters.native import dbsql_parameter_from_primitive
 
-from databricks.sql.utils import ParamEscaper, inject_parameters, transform_paramstyle, ParameterStructure
+from databricks.sql.parameters.inline import ParamEscaper, inject_parameters, transform_paramstyle, ParameterStructure
 
 pe = ParamEscaper()
